Add newline above .. code:: directive
RST doesn't recognize .. code:: or any other directive unless it is on a
new paragraph. This means you need to make sure there's a blank line
above where you generate it.

This fixes #1 by simply adding a newline above the generated code block.
In the future, it would be better to keep track of whether we're on a
new paragraph, so a newline isn't unnecessarily added.